refactor: move configuration ui primitives out of shared
