--- src/renderer/shared/ui/settings/SettingsPrimitives.tsx
@@ -1,5 +1,0 @@
-export { SettingsActionCard, SettingsSectionCard, SettingsStatTile, SettingsStatusBadge, SettingsSurfaceCard, SettingsToggleCard } from './SettingsPrimitives.cards';
-export { SettingsField, SettingsSelectField, SettingsTextAreaField } from './SettingsPrimitives.fields';
-export { SettingsNotice } from './SettingsPrimitives.feedback';
-export { SettingsModal } from './SettingsPrimitives.modal';
-export { settingsButtonClassName } from './SettingsPrimitives.shared';
